refactor: split gym protocol module
